+++ src/id_translation/fetching/_multi_fetcher.py
@@ -43,7 +43,7 @@ class MultiFetcher(Fetcher[SourceType, IdType]):
         *fetchers: Fetcher[SourceType, IdType],
         max_workers: int = 2,
         duplicate_translation_action: ActionLevel = ActionLevel.WARN,
-        duplicate_source_discovered_action: ActionLevel = ActionLevel.IGNORE,
+        duplicate_source_discovered_action: ActionLevel = ActionLevel.WARN,
     ) -> None:
         for pos, f in enumerate(fetchers):
             if not isinstance(f, Fetcher):  # pragma: no cover